'add_multithread'

+++ post.py
@@ -5,6 +5,7 @@ from flask_restx import Api, Namespace, Resource, fields
 from config import DB
 from werkzeug.utils import secure_filename
 from auth_util import login_required, check_access_token
+from mysql.connector import pooling
 
 conn = pymysql.connect(
         host=DB['host'], 
@@ -13,6 +14,18 @@ conn = pymysql.connect(
         password=DB['user_pw'], 
         database=DB['database'], 
         charset=DB['charset'])
+
+connection_pool = pooling.MySQLConnectionPool(
+    pool_name = "capstone_pool",
+    pool_size = 3,
+    pool_reset_session = True,
+    host = DB['host'],
+    port = DB['port'],
+    user = DB['user_id'],
+    password = DB['user_pw'],
+    database = DB['database'],
+    charset = DB['charset']
+)
 
 Post_api = Namespace(
     name="Post",
@@ -49,17 +62,18 @@ class add_post(Resource):
             start_date = request.json['start_date']
             end_date = request.json['end_date']
 
-            db_conn = conn
-            with db_conn.cursor(pymysql.cursors.DictCursor) as cursor:
-                cursor.execute(f"INSERT INTO post (title, contacts, email, assignee, registration_method, address, detailed_link, start_date, end_date, registration_date) VALUES ('{title}', '{contacts}', '{email}', '{assignee}', '{registration_method}', '{address}', '{detailed_link}','{start_date}', '{end_date}', DATE_FORMAT(now() + interval 9 hour, '%Y-%m-%d %H:%i'))")
-
-            db_conn.commit()
+            connection_obj = connection_pool.get_connection()
+            if connection_obj.is_connected():
+                with connection_obj.cursor(pymysql.cursors.DictCursor) as cursor:
+                    sql = f"INSERT INTO post (title, contacts, email, assignee, registration_method, address, detailed_link, start_date, end_date, registration_date) VALUES ('{title}', '{contacts}', '{email}', '{assignee}', '{registration_method}', '{address}', '{detailed_link}','{start_date}', '{end_date}', DATE_FORMAT(now() + interval 9 hour, '%Y-%m-%d %H:%i'))"
+                    cursor.execute(sql)
+                    cursor.commit()
+                    connection_obj.close()
             return {
                 "message" : "success"
             },200
         
         except Exception as e:
-            db_conn.close()
             return str(e)
         
 @Post_api.route('/get_post', methods = ['GET'])
@@ -68,11 +82,14 @@ class get_post(Resource):
     @login_required
     def get(self):
         try:
-            db_conn = conn
-            with db_conn.cursor(pymysql.cursors.DictCursor) as cursor:
-                cursor.execute("SELECT * FROM post")
-                posts = cursor.fetchall()
-            return {"post_items" : posts}
+            connection_obj = connection_pool.get_connection()
+            if connection_obj.is_connected():
+                with connection_obj.cursor(pymysql.cursors.DictCursor) as cursor:
+                    sql = "SELECT * FROM post"
+                    cursor.execute(sql)
+                    posts = cursor.fetchall()
+                    connection_obj.close()
+                return {"post_items" : posts}
         except Exception as e:
             return str(e)
         finally:
@@ -84,10 +101,13 @@ class get_post(Resource):
     @login_required
     def get(self, post_id):
         try:
-            db_conn = conn
-            with db_conn.cursor(pymysql.cursors.DictCursor) as cursor:
-                cursor.execute(f"SELECT * FROM post WHERE post_id = {post_id}")
-                posts = cursor.fetchone()
+            connection_obj = connection_pool.get_connection()
+            if connection_obj.is_connected():
+                with connection_obj.cursor(pymysql.cursors.DictCursor) as cursor:
+                    sql = f"SELECT * FROM post WHERE post_id = {post_id}"
+                    cursor.execute(sql)
+                    posts = cursor.fetchone()
+                    connection_obj.close()
             return posts
         except Exception as e:
             return str(e)
@@ -100,19 +120,22 @@ class delete_post(Resource):
     @login_required
     def delete(self, post_id):
         try:
-            db_conn = conn
-            with db_conn.cursor(pymysql.cursors.DictCursor) as cursor:
-
-                cursor.execute(f"SELECT * FROM post WHERE post_id = {post_id}")
-                exist = cursor.fetchone()
+            connection_obj = connection_pool.get_connection()
+            if connection_obj.is_connected():
+                with connection_obj.cursor(pymysql.cursors.DictCursor) as cursor:
+                    sql = f"SELECT * FROM post WHERE post_id = {post_id}"
+                    cursor.execute(sql)
+                    exist = cursor.fetchone()
 
                 if exist is None:
+                    connection_obj.close()
                     return {
                         "message" : "fail"
                     },200
                 else:
                     cursor.execute(f"DELETE FROM post WHERE post_id = {post_id}")
-                    conn.commit()
+                    connection_obj.commit()
+                    connection_obj.close()
                     return {
                         "message" : "success"
                     },200
@@ -129,10 +152,12 @@ class update_post(Resource):
         try:
             title = request.json['title']
             content = request.json['content']
-            db_conn = conn
-            with db_conn.cursor(pymysql.cursors.DictCursor) as cursor:
-                cursor.execute(f"UPDATE post SET title = '{title}', content = '{content}' WHERE post_id = {post_id}")
-                conn.commit()
+            connection_obj = connection_pool.get_connection()
+            if connection_obj.is_connected():
+                with connection_obj.cursor(pymysql.cursors.DictCursor) as cursor:
+                    cursor.execute(f"UPDATE post SET title = '{title}', content = '{content}' WHERE post_id = {post_id}")
+                    connection_obj.commit()
+                    connection_obj.close()
                 return {
                     "message" : "success"
                 },200
